requirement.txt
openai>=1.12.0
python-dotenv>=1.0.0
requests>=2.31.0

msal>=1.26.0

google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.122.0pip install --upgrade --force-reinstall -r requirement.txt